requirements.txt
groq
python-dotenv
streamlit
PyPDF2
python-docx